[Refactor] Move JSON functions into ExprCore
