+++ be/src/exprs/hyper_json_transformer.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "util/hyper_json_transformer.h"
+#include "exprs/hyper_json_transformer.h"
 
 #include <cstddef>
 #include <memory>
