+++ be/src/exprs/json_functions.cpp
@@ -40,15 +40,14 @@
 #include "exprs/column_ref.h"
 #include "exprs/function_context.h"
 #include "exprs/function_helper.h"
+#include "exprs/hyper_json_transformer.h"
 #include "exprs/jsonpath.h"
 #include "glog/logging.h"
 #include "gutil/casts.h"
 #include "gutil/strings/substitute.h"
-#include "storage/chunk_helper.h"
 #include "types/json_value.h"
 #include "types/logical_type.h"
 #include "types/type_descriptor.h"
-#include "util/hyper_json_transformer.h"
 #include "velocypack/Builder.h"
 #include "velocypack/Iterator.h"
