+++ be/src/storage/rowset/json_column_compactor.cpp
@@ -26,12 +26,12 @@
 #include "column/flat_json/json_merger.h"
 #include "column/json_column.h"
 #include "column/nullable_column.h"
+#include "exprs/hyper_json_transformer.h"
 #include "gen_cpp/segment.pb.h"
 #include "gutil/casts.h"
 #include "storage/flat_json_metrics.h"
 #include "storage/rowset/column_writer.h"
 #include "types/constexpr.h"
-#include "util/hyper_json_transformer.h"
 #include "util/json_path_deriver.h"
 
 namespace starrocks {
