+++ be/test/CMakeLists.txt
@@ -162,7 +162,6 @@ set(EXEC_FILES
         ./exprs/function_call_expr_test.cpp
         ./exprs/geography_functions_test.cpp
         ./exprs/if_expr_test.cpp
-        ./exprs/json_functions_test.cpp
         ./exprs/flat_json_functions_test.cpp
         ./exprs/lambda_array_expr_test.cpp
         ./exprs/lambda_map_expr_test.cpp
