+++ be/test/exprs/CMakeLists.txt
@@ -38,6 +38,7 @@ set(EXPR_CORE_TEST_FILES
     es_functions_test.cpp
     function_helper_test.cpp
     gin_functions_test.cpp
+    json_functions_test.cpp
     map_expr_test.cpp
     map_element_expr_test.cpp
     map_functions_test.cpp
